mvc version
The host has version 3.0.0.0 and the application used 3.0.0.1

In order to get the application to run on the host I had to update the
configured value in both the root and view web.config files to 3.0.0.1
and had to mark mvc.dll to copy local true. I copied the dll files out
to server bin folder.